+++ release_notes.md
@@ -1,11 +1,37 @@
 # **Release notes**
 
+## **Release Notes develop**
+
+* Added camera config script == videopc
+* Added CI builds for windows/ubuntu install
+* Optimized and fixed broken install procedure
+* Added script for calculation of wheel to screen positions
+* Added script for getting screen positions from wheel input (postprocessing)
+* Separated local param methods from Alyx board methods usning alyx module
+* Added tests for param module and path helper object
+* Created method for finding mapped network drives on local rigs
+* Added local and remote data folder to params_alyx
+* Swapped training camera timestamps/bonsai timestamps
+* Added backup params file on write
+* Added register_screen_lux script
+* Transfer data will attempt to move all passive session that it finds before starting
+* Fixed stimulus overshooting/undershooting on center or right/left thresholds
+* Updated Bonsai and packages
+* Added stimulus capability for reverse contingencies
+* Updated Bonsai deployment methods
+* Removed all ONE calls from task launching procedures
+* Added sonic studio config file for xonar sound card config (in iblrig/devices/sound_card/IBL.nsx2)
+* HabituationCW now uses IBL default stimulus
+* Fixed bug in camera recording where SAVE_VIDEO = False would make stream not start (deprecated task settings SAVE_VIDEO flag)
+* Added bonsai.show_stim method to BpodMessageCreator obj
+* Deprecated RECORD_VIDEO "knob"
+
 ## **Release Notes 6.4.2**
 
 Patch update (bugfixes)
 
 * Increased stim_off state timer to 150ms
-* Updated ibllib, Bonsai and Bonvision 
+* Updated ibllib, Bonsai and Bonvision
 * New update procedure for Bonsai that will simplify updates for users
 * New pipeline architecture (removed deprecated flag creations on transfer)
 * Mice that have more than one project now will require user to pick the project on run
@@ -48,12 +74,12 @@ Minor update (added functionality)
 ## **Release Notes 6.2.5**
 
 *THIS: State Machine changed*
-Patch update (bugfixes)  
+Patch update (bugfixes)
 
 * Fixed stimulus sometimes keeps moving after reward for all tasks
 * Fixed session_params bug in ephysCW
 * Under the hood refactorings
-  
+
 ## **Release Notes 6.2.4**
 
 Patch update (bugfixes)
@@ -73,13 +99,13 @@ Patch update (bugfixes)
 
 ## **Release Notes 6.2.2**
 
-Patch update (bugfixes)  
+Patch update (bugfixes)
 
 * SESSION_ORDER bugfix
 
 ## **Release Notes 6.2.1**
 
-Patch update (bugfixes)  
+Patch update (bugfixes)
 Mainly in **ephysCW** and **passiveCW**
 
 * Poop only at end o passive run